+++ zombie_game/Floor.h
@@ -7,7 +7,7 @@
 class Floor : public Static
 {
 	public:
-		Floor(AssetManager& assets, sf::Vector2f pos, sf::Vector2f size);
+		Floor(AssetManager& assets, sf::Vector2f pos, sf::Vector2i size);
 };
 
 #endif
