+++ zombie_game/Wall.cpp
@@ -1,6 +1,6 @@
 #include "Wall.h"
 
-Wall::Wall(AssetManager& assets, sf::Vector2f pos, sf::Vector2f size) : Static(assets)
+Wall::Wall(AssetManager& assets, sf::Vector2f pos, sf::Vector2i size) : Static(assets)
 {
 	this->sprite.setTexture(assets.wallTexture);
 	this->sprite.setTextureRect(sf::IntRect(0, 0, size.x, size.y));
